refactor(integrations): flatten structure\n\n- Keep only integrations/index.ts at top level\n- Move notion.ts, slack.ts, discord.ts into integrations/providers/\n- Introduce integrations/core/ for shared logic: util.ts, composio.ts, integration-sync.ts\n- Update imports and route sync import accordingly\n- Create branch refactor/integrations-structure
